pyproject.toml
[project]
name = "fec"
version = "0.1.0"
description = "Federal Election Commission data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.10.13",
    "dbt-snowflake>=1.10.2",
    "duckdb>=1.4.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "snowflake-connector-python>=3.18.0",
    "pyarrow>=19.0.0",
    "pandas>=2.0.0",
]
